fixing warning, refs #3
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@23263 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/utils/geom/PositionVector.cpp
@@ -1258,7 +1258,7 @@ PositionVector::getOrthogonal(const Position& p, double extend, bool before, dou
         }
     } else {
         // take the segment after closestIndex if possible
-        if (closestIndex < (size() - 1)) {
+        if (closestIndex < (int)size() - 1) {
             result.push_back(tmp[closestIndex + 1]);
         } else {
             result.push_back(tmp[-1]);
